Change babel presets from es2015 to env | Got it workin
